http://issues.apache.org/activemq/browse/AMQCPP-103
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@617637 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/io/LoggingInputStream.cpp
@@ -65,7 +65,7 @@ int LoggingInputStream::read( unsigned char* buffer,
 
         log( buffer, numRead );
 
-        return numRead;
+        return (int)numRead;
     }
     AMQ_CATCH_RETHROW( IOException )
     AMQ_CATCHALL_THROW( IOException )
